Fix: move supplier order creation to make space for supplier/id mgmt view
